[S-104746] Script to roll dates forward for demo/training
This is a script that will move all dates forward by a set
number of days, 1 by default.

It has been changed so that it will not move changes into the future.
Instead, compress the audit records that would be moved to the future
into the 30 seconds before now. This will preserve the property of the
Audit table that increasing Audit IDs have increasing Audit ChangeDateUTC
values when subsequent updates are made by users to the Agility Core
instance.

The query generation code in Agility Core makes the assumption that it
can "select a moment slice", i.e. it can select an Audit ID range and
get all the states in between two dates. If the above property is not
preserved, the "moment slice" is not guaranteed to be correct.